+++ package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "@vercel/speed-insights": "^1.0.9",
-    "moorhen": "0.8.5",
+    "moorhen": "0.9.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^9.0.4",
